+++ packages/client/src/subscribe/Subscription.ts
@@ -28,11 +28,14 @@ export class Subscription extends MessageStream {
     protected readonly logger: Logger
     readonly streamPartId: StreamPartID
     protected eventEmitter: EventEmitter<SubscriptionEvents>
+    /** @internal */
+    readonly isRaw: boolean
 
     /** @internal */
-    constructor(streamPartId: StreamPartID, loggerFactory: LoggerFactory) {
+    constructor(streamPartId: StreamPartID, isRaw: boolean, loggerFactory: LoggerFactory) {
         super()
         this.streamPartId = streamPartId
+        this.isRaw = isRaw
         this.eventEmitter = new EventEmitter<SubscriptionEvents>()
         this.logger = loggerFactory.createLogger(module)
         this.onMessage.listen((msg) => {