+++ packages/dht/test/integration/ScaleDownDht.test.ts
@@ -3,7 +3,7 @@ import { DhtNode } from '../../src/dht/DhtNode'
 import { NodeType, PeerDescriptor } from '../../src/proto/packages/dht/protos/DhtRpc'
 import { createMockConnectionDhtNode } from '../utils/utils'
 import { isSamePeerDescriptor } from '../../src/helpers/peerIdFromPeerDescriptor'
-import { Logger, waitForCondition } from '@streamr/utils'
+import { Logger } from '@streamr/utils'
 
 const logger = new Logger(module)
 
@@ -12,7 +12,7 @@ describe('Scaling down a Dht network', () => {
     let nodes: DhtNode[]
     let entrypointDescriptor: PeerDescriptor
     const simulator = new Simulator(LatencyType.RANDOM)
-    const NUM_NODES = 42
+    const NUM_NODES = 80
     const MAX_CONNECTIONS = 15
     const K = 2
 
@@ -48,33 +48,21 @@ describe('Scaling down a Dht network', () => {
         for (let i = 1; i < nodes.length; i++) {
             randomIndices.push(i)
         }
-        let failedCleanUps = 0
         while (randomIndices.length > 1) {
             const index = Math.floor(Math.random() * randomIndices.length)
             const nodeIndex = randomIndices[index]
             randomIndices.splice(index, 1)
             const stoppingPeerDescriptor = nodes[nodeIndex].getPeerDescriptor()
             await nodes[nodeIndex].stop()
-            try {
-                await waitForCondition(() =>
-                    nodes.every((node) =>
-                        node.getAllConnectionPeerDescriptors().every((peer) => {
-                            if (isSamePeerDescriptor(peer, stoppingPeerDescriptor)) {
-                                logger.trace(' ' + node.getNodeName() + ', ' + stoppingPeerDescriptor.nodeName + ' cleaning up failed')
-                            }
-                            return !isSamePeerDescriptor(peer, stoppingPeerDescriptor)
-                        })
-                    )
-                )
-            } catch (err) {
-                const failures = nodes.reduce((total, node) =>
-                    total + node.getAllConnectionPeerDescriptors().reduce((acc, peer) =>
-                        isSamePeerDescriptor(peer, stoppingPeerDescriptor) ? acc + 1 : acc
-                    , 0)
-                , 0)
-                failedCleanUps += failures
-            }
-            expect(failedCleanUps).toBeLessThan(1)
+            const nodeIsCleaned = nodes.every((node) =>
+                node.getAllConnectionPeerDescriptors().every((peer) => {
+                    if (isSamePeerDescriptor(peer, stoppingPeerDescriptor)) {
+                        logger.error(' ' + node.getNodeName() + ', ' + stoppingPeerDescriptor.nodeName + ' cleaning up failed')
+                    }
+                    return !isSamePeerDescriptor(peer, stoppingPeerDescriptor)
+                })
+            )
+            expect(nodeIsCleaned).toEqual(true)
         }
     }, 180000)
 })