+++ packages/dht/test/integration/WebRtcConnectionManagement.test.ts
@@ -46,10 +46,12 @@ describe('WebRTC Connection Management', () => {
     })
 
     afterEach(async () => {
-        await manager1.stop()
-        await manager2.stop()
-        await connectorTransport1.stop()
-        await connectorTransport2.stop()
+        await Promise.all([
+            manager1.stop(),
+            manager2.stop(),
+            connectorTransport1.stop(),
+            connectorTransport2.stop()
+        ])
         simulator.stop()
     })